pyproject.toml
[tool.poetry]
name = "test-task-drf"
version = "0.1.0"
description = ""
authors = ["moonvent <terraboneystone@proton.me>"]
readme = "README.md"
packages = [{include = "test_task_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"
loguru = "^0.6.0"
debugpy = "^1.6.6"
python-dotenv = "^0.21.1"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
drf-yasg = "^1.21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
